Better abstraction nodes.
